+++ PotteryBarn/Core/Pieces/VanillaShop.cs
@@ -1,0 +1,12 @@
+﻿namespace PotteryBarn;
+
+using static PieceConstants;
+
+public static class VanillaShop {
+  public static readonly PotteryPieceList CultivatorPieces =
+      new PotteryPieceList()
+          .AddPiece(
+              "vines",
+              Workstation.None,
+              [ (Item.Wood, 2) ]);
+}